rename Feedstock page and store components to use slug instead of id for improved routing
