+++ app/feedstock/[slug]/code-snippet.tsx
