+++ app/feedstock/[slug]/data-viewer.tsx
