+++ app/feedstock/[slug]/details.tsx
