+++ app/feedstock/[slug]/header.tsx
